feat: atualiza tipo da média para Double na classe Livro e implementa PerfilRequestDTO
Este commit implementa melhorias na entidade `Livro` e na funcionalidade de criação e atualização de perfis. Agora, a média do livro foi alterada para o tipo `Double`, e um novo DTO (`PerfilRequestDTO`) foi criado para melhorar a entrada de dados no sistema. Além disso, foram feitas atualizações na documentação das classes `Usuario` e `PerfilService`, e ajustes nas permissões de usuário no `SecurityConfig` para permitir que usuários comuns possam cadastrar livros.

- Melhorias na entidade `Livro`:
  - Altera o tipo do campo `media_livro` para `Double`, permitindo valores decimais na média do livro.

- Implementa `PerfilRequestDTO`:
  - Cria um novo DTO para representar a entrada de dados ao criar ou atualizar um perfil.
  - Atualiza os métodos `savePerfil` e `updatePerfil` no `PerfilService` e `PerfilServiceImpl` para utilizar o novo DTO.
  - Melhora a validação e a documentação dos campos no `PerfilRequestDTO`.

- Atualizações na documentação:
  - Atualiza a documentação da classe `Usuario` para refletir as mudanças nos campos e regras de negócio.
  - Atualiza a documentação do `PerfilService` para incluir detalhes sobre o uso do `PerfilRequestDTO`.

- Ajustes no `SecurityConfig`:
  - Permite que usuários comuns possam cadastrar livros, ajustando as permissões de acesso.

- Outras mudanças:
  - Refatora o `PerfilController` para utilizar o `PerfilRequestDTO` nas operações de criação e atualização de perfis.

+++ src/main/java/prati/projeto/redeSocial/modal/entity/Livro.java
@@ -85,5 +85,5 @@ public class Livro {
 
     @Column(name = "media_livro")
     @Schema(description = "Média com base na avaliação do livro")
-    private Integer media = 0;
+    private Double media = 0.0;
 }
